second commit

+++ Algo/algo.go
@@ -65,7 +65,7 @@ func EqualizeArray(arr []int32) int32 {
 
 	}
 	
-	return int32(del)
+	return del
 
 }
